refactor: move file to top-dir
